* commit 'fe325c16fa4054983cf3013f9fd7276ea86294e7 [formerly f84dca53480b17b8e67b71b8f9d65acae7c35d86]': (148 commits)
  DF-10716 fix one more time cmake build for win32
  DF-10716 fix asset-server build
  DF-10716 fix cmake dava-template var
  DF-10716 try fix mounting for dependent packs (for Denis, need test. TODO)
  DF-10716 second fix win32 cmake config
  DF-10716 fix cmake config for mac or win
  DF-10716 try fix TestBed package for wint32 (add scripts dir to archive)
  DF-10716 fix UAP build
  DF-10716 fix crush in CurDownloader uint32 -> long 8 bytes on mac os
  DF-10716 fix test bed local server start stop with Terminal and quit from terminal
  DF-10716 test bed works with local python http server (tested on win32)
  DF-10716 fix unit test build
  DF-10716 start fixing unit tests
  DF-10716 fix test_bed pack_manager downloading and building packs
  DF-10716 continue fixing test_bed (not change pack_mananger)
  DF-10716 fix build with trunk client
  DF-10716 change url to server start fixing test bed for new pack generation
  DF-10716 fix not UNITY build after merge
  DF-10716 fix comments review
  [DF-10716] reverting back %<group> usage in commands due to random crashes
  ...


Former-commit-id: 240b5c09c766cd9fb893c023c36aa363c0b58a2f

+++ Sources/Internal/Render/Image/LibJpegHelper.cpp
@@ -65,7 +65,7 @@ eErrorCode LibJpegHelper::ReadFile(File* infile, Vector<Image*>& imageSet, const
     jpegErrorManager jerr;
 
     infile->Seek(0, File::SEEK_FROM_START);
-    uint32 fileSize = infile->GetSize();
+    uint32 fileSize = static_cast<uint32>(infile->GetSize());
     uint8* fileBuffer = new uint8[fileSize];
     infile->Read(fileBuffer, fileSize);
     infile->Seek(0, File::SEEK_FROM_START);
@@ -99,7 +99,7 @@ eErrorCode LibJpegHelper::ReadFile(File* infile, Vector<Image*>& imageSet, const
 #endif
 
     jpeg_create_decompress(&cinfo);
-    
+
 #ifdef __clang__
 #pragma clang diagnostic pop
 #endif
@@ -220,14 +220,14 @@ eErrorCode LibJpegHelper::WriteFile(const FilePath& fileName, const Vector<Image
         SafeRelease(convertedImage);
         return eErrorCode::ERROR_WRITE_FAIL;
     }
-    
+
 #ifdef __clang__
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wold-style-cast"
 #endif
 
     jpeg_create_compress(&cinfo);
-    
+
 #ifdef __clang__
 #pragma clang diagnostic pop
 #endif
@@ -277,7 +277,7 @@ DAVA::ImageInfo LibJpegHelper::GetImageInfo(File* infile) const
     DVASSERT(infile);
 
     infile->Seek(0, File::SEEK_FROM_START);
-    uint32 fileSize = infile->GetSize();
+    uint32 fileSize = static_cast<uint32>(infile->GetSize());
     uint8* fileBuffer = new uint8[fileSize];
     infile->Read(fileBuffer, fileSize);
     cinfo.err = jpeg_std_error(&jerr.pub);
@@ -290,7 +290,7 @@ DAVA::ImageInfo LibJpegHelper::GetImageInfo(File* infile) const
         infile->Seek(0, File::SEEK_FROM_START);
         return ImageInfo();
     }
-    
+
 #ifdef __clang__
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wold-style-cast"
